requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==2.2.2
django-cors-headers==3.5.0
djangorestframework==3.11.1
gTTS==2.1.1
gTTS-token==1.1.3
gunicorn==20.0.4
idna==2.10
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
